Moved publishing specific code to its own file.
